feature: crew alert 및 헤더 js 추가

+++ src/main/webapp/crew/detail.jsp
@@ -9,11 +9,13 @@
 <meta charset="UTF-8">
 <title>MEETUP</title>
 <!-- CSS -->
-<link rel="stylesheet" href="./css/reset.css"  type="text/css" />
-<link rel="stylesheet" href="./css/index.css"  type="text/css" />
+<link rel="stylesheet" href="./css/reset.css" type="text/css" />
+<link rel="stylesheet" href="./css/index.css" type="text/css" />
+<link rel="stylesheet" href="./css/header.css" type="text/css" />
 <!-- CDN -->
-<!-- <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script> -->
+<script src="component/header.js"></script>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/sweetalert2@10"></script>
 <script src="https://unpkg.com/boxicons@2.1.4/dist/boxicons.js"></script>
 <style>
 	.container {
@@ -105,6 +107,7 @@
         font-size: 14px;
         width: 650px;
         color: white;
+        text-align: center;
 	}
 	.like-icon {
 		height: 30px;
@@ -178,37 +181,6 @@
 		color: black;
 		margin-bottom: 5px;
 	}
-	/* 모달 */
-	/* .modal {
-		display: none; 
-	    position: absolute;
-	    top:0;
-	    left: 0;
-	    width: 100%;
-	    height: 100vh;
-	    overflow: hidden;
-	    background: rgba(0,0,0,0.5);
-	}
-	.modal .modal_popup {
-	    position: absolute;
-	    top: 50%;
-	    left: 50%;
-	    transform: translate(-50%, -50%);
-	    padding: 20px;
-	    background: #ffffff;
-	    border-radius: 20px;
-	}
-	.modal .modal_popup .close_btn {
-	    display: block;
-	    padding: 10px 20px;
-	    background-color: rgb(116, 0, 0);
-	    border: none;
-	    border-radius: 5px;
-	    color: #fff;s
-	    cursor: pointer;
-	    transition: box-shadow 0.2s;
-	} */
-	
 	/* 정모 박스 */
 	.meetingOne {
 	  width: 50%;
@@ -269,11 +241,11 @@
 			</c:when>
 	        <c:when test="${role eq 'crewMember' || role eq 'adminMember'}">
 				<a href="#">신고하기</a>
-	        		<a href="crew?cmd=mng&requestType=leave&crewNo=${crewVO.crewNo}&memberNo=${sessionScope.loginMember.memberNo}">모임퇴장</a>
+	        		<a href="crew?cmd=mng&requestType=leave&crewNo=${crewVO.crewNo}&memberNo=${sessionScope.loginMember.memberNo}" class="crew-alert-confirm" data-message="모임을 퇴장하시겠습니까?" >모임퇴장</a>
 			</c:when>
 			<c:otherwise>
-				<a href="crew?cmd=update&crewNo=${crewVO.crewNo}">수정하기</a>
-	        		<a href="crew?cmd=delete&crewNo=${crewVO.crewNo}">삭제하기</a>
+				<a href="crew?cmd=update&crewNo=${crewVO.crewNo}" >수정하기</a>
+	        		<a href="crew?cmd=delete&crewNo=${crewVO.crewNo}" class="crew-alert-confirm" data-message="모임을 삭제하시겠습니까?" >삭제하기</a>
 			</c:otherwise>
 			</c:choose>
 	    </div>
@@ -290,31 +262,31 @@
 						<a href="member?cmd=login" class="btn btn-main">로그인하고 모임 가입하기</a>
 					</c:when>
 					<c:when test="${role eq 'member'}">
-						<a href="crew?cmd=signup&crewNo=${crewVO.crewNo}" class="btn btn-main">가입신청</a>
+						<a href="crew?cmd=signup&crewNo=${crewVO.crewNo}" class="btn btn-main crew-alert" data-message="가입신청 되었습니다." >가입신청</a>
 					</c:when>
 					<c:when test="${role eq 'pendingMember'}">
-						<a href="" class="btn btn-main">승인중</a>
+						<div class="btn btn-main">승인중</div>
 					</c:when>
 					<c:when test="${role eq 'rejectedMember'}">
-						<a href="" class="btn btn-main">승인 거절</a>
+						<div class="btn btn-main">승인 거절된 모임입니다</div>
 					</c:when>
 					<c:when test="${role eq 'kick'}">
-						<a href="" class="btn btn-main">강퇴된 모임</a>
+						<div class="btn btn-main">강퇴된 모임입니다</div>
 					</c:when>
 					<c:otherwise>
-						<a href="" class="btn btn-main">공유하기</a>
+						<a href="" class="btn btn-main">공유</a>
 					</c:otherwise>
 				</c:choose>
 				<c:if test="${isValidCrewLike eq 0}">
 					<c:if test="${empty sessionScope.loginMember}">
-						<a href="member?cmd=login"><box-icon class="like-icon" name='heart' ></box-icon></a>
+						<a href="member?cmd=login" class="crew-alert" data-message="로그인 후 이용 가능합니다." ><box-icon class="like-icon" name='heart' ></box-icon></a>
 					</c:if>
 					<c:if test="${not empty sessionScope.loginMember}">
-						<a href="crew?cmd=like&requestType=add&crewNo=${crewVO.crewNo}"><box-icon class="like-icon" name='heart' ></box-icon></a>
+						<a href="crew?cmd=like&requestType=add&crewNo=${crewVO.crewNo}" ><box-icon class="like-icon" name='heart' ></box-icon></a>
 					</c:if>
 				</c:if>
 				<c:if test="${isValidCrewLike eq 1}">
-					<a href="crew?cmd=like&requestType=delete&crewNo=${crewVO.crewNo}"><box-icon class="like-icon" name='heart' type='solid' ></box-icon></a>
+					<a href="crew?cmd=like&requestType=delete&crewNo=${crewVO.crewNo}" ><box-icon class="like-icon" name='heart' type='solid' ></box-icon></a>
 				</c:if>
 			</div>
 		</div>
@@ -349,124 +321,84 @@
 	<!-- 정모 -->
 	<div class="meeting-container">
 		<p class="meeting-logo">정모</p>
-		<c:if test="${role eq 'leader' || role eq 'adminMember'}">
+		<c:if test="${role eq 'crewMember' || role eq 'adminMember'}">
 			<a href="meeting?cmd=write&crewNo=${crewVO.crewNo}">정모 등록하기</a>
 		</c:if>
 		
 		<div>
 		
-		<c:forEach var="meetingVO" items="${meetingList}" varStatus="status">
-			<c:set var="today" value="<%=new java.util.Date() %>"/>
-			<fmt:formatDate var="nowDay" value="${today}" pattern="yyyy-MM-dd hh:mm:ss" />
-			<fmt:parseDate value="${nowDay}" var="pdToday" pattern="yyyy-MM-dd"/>
-			<fmt:parseNumber value="${pdToday.time / (1000*60*60*24)}" integerOnly="true" var="numToday"></fmt:parseNumber>
-			
-			<c:set var="meetingDate" value="${meetingVO.meetingDate}"/>
-			<fmt:parseDate value="${meetingDate}" var="pdMeetingDate" pattern="yyyy-MM-dd"/>
-			<fmt:parseNumber value="${pdMeetingDate.time / (1000*60*60*24)}" integerOnly="true" var="numMeetingDate"></fmt:parseNumber>
-			<c:set var="isMeetingMember" value="false" />
-			
-			<!-- 정모와 정모참여 정모no가 같은지 확인, 정모에 참여한 멤버가 로그인한 유저인지 확인 -->
-			<c:forEach var="meetingMemberVO" items="${meetingMemberList}">
-				<c:if test="${meetingMemberVO.memberNo == MemberVO.memberNo && meetingVO.meetingNo eq meetingMemberVO.meetingNo}">
-						<c:set var="isMeetingMember" value="true" />
-				</c:if>
-			</c:forEach>
-			
-			<!-- for문을 돌면서 정모정보를 계속 뿌려주는 div -->
-			<div class="meetingOne">
-				<div style="display: inline-block">
-					<span><fmt:formatDate value="${meetingDate}" pattern="MM/dd (E)"/></span> 
-					<span style=color:red>D-${numMeetingDate - numToday}</span>
-					<p>${meetingVO.meetingName}</p>
-					<c:if test="${role eq 'crewMember' || role eq 'adminMember' || role eq 'leader'}">
-					<c:choose>
-						<c:when test="${not empty meetingMemberList}">
-							<c:choose>
-								<c:when test="${isMeetingMember}" >
-									<span><a href="meetingExit?memberNo=${MemberVO.memberNo}&meetingNo=${meetingVO.meetingNo}&crewNo=${crewVO.crewNo}">정모 나가기</a></span>
-				        		</c:when>
-				        		<c:otherwise>
-				        			<span><a href="meetingAttend?memberNo=${MemberVO.memberNo}&meetingNo=${meetingVO.meetingNo}&crewNo=${crewVO.crewNo}">정모 참석하기</a></span>
-				        		</c:otherwise>      	
-		        			</c:choose>
-		        		</c:when>
-		        		<c:otherwise>
-		        			<span><a href="meetingAttend?memberNo=${MemberVO.memberNo}&meetingNo=${meetingVO.meetingNo}&crewNo=${crewVO.crewNo}">정모 참석하기</a></span>
-		        		</c:otherwise>
-	        		</c:choose>
-					</c:if>
-				</div>
+		<c:forEach var="meetingVO" items="${meetingList}">
+				<c:set var="today" value="<%=new java.util.Date() %>"/>
+				<fmt:formatDate var="nowDay" value="${today}" pattern="yyyy-MM-dd hh:mm:ss" />
+				<fmt:parseDate value="${nowDay}" var="pdToday" pattern="yyyy-MM-dd"/>
+				<fmt:parseNumber value="${pdToday.time / (1000*60*60*24)}" integerOnly="true" var="numToday"></fmt:parseNumber>
+				
+				<c:set var="meetingDate" value="${meetingVO.meetingDate}"/>
+				<fmt:parseDate value="${meetingDate}" var="pdMeetingDate" pattern="yyyy-MM-dd"/>
+				<fmt:parseNumber value="${pdMeetingDate.time / (1000*60*60*24)}" integerOnly="true" var="numMeetingDate"></fmt:parseNumber>
+				
+				<div class="meetingOne">
+					<div style="display: inline-block">
+						<span><fmt:formatDate value="${meetingDate}" pattern="MM/dd (E)"/></span> 
+						<span style=color:red>D-${numMeetingDate - numToday}</span>
+						<p>${meetingVO.meetingName}</p>
+						<c:if test="${role eq 'crewMember' || role eq 'adminMember'}">
+						<c:choose>
+							<c:when test="${not empty meetingMemberList}">
+								<h2>${meetingMemberList[status.index].meetingNo}</h2>
+								<h2>${meetingVO.meetingNo}</h2>
+									<c:choose>
+										<c:when test="${meetingMemberList[status.index].meetingNo eq meetingVO.meetingNo}" >
+											<span><a href="meetingExit?memberNo=${MemberVO.memberNo}&meetingNo=${meetingVO.meetingNo}&crewNo=${crewVO.crewNo}">정모 나가기</a></span>
+						        		</c:when>
+						        		<c:otherwise>
+						        			<span><a href="meetingAttend?memberNo=${MemberVO.memberNo}&meetingNo=${meetingVO.meetingNo}&crewNo=${crewVO.crewNo}">정모 참석하기</a></span>
+						        		</c:otherwise>      	
+				        			</c:choose>
+			        		</c:when>
+			        		<c:otherwise>
+			        			<span><a href="meetingAttend?memberNo=${MemberVO.memberNo}&meetingNo=${meetingVO.meetingNo}&crewNo=${crewVO.crewNo}">정모 참석하기</a></span>
+			        		</c:otherwise>
+		        		</c:choose>
+						</c:if>
+					</div>
 
-				<br />
-				<div class="meeting-info">
-					<img style="width: 100px; height:100px;" src="upload/${meetingVO.meetingSaveImg}" alt="" />
-					<div>
-						<div>일시 <fmt:formatDate value="${meetingDate}" pattern="MM/dd (E) aa hh:mm "/></div>
-						<div>위치 ${meetingVO.meetingLoc}</div>
-						<div>비용 ${meetingVO.meetingPrice}원</div>
-						<div>참석 ${meetingVO.meetingAttend}/${meetingVO.meetingMax} (${meetingVO.meetingMax-meetingVO.meetingAttend}명남음) </div>
+					<br />
+					<div class="meeting-info">
+						<img style="width: 100px; height:100px;" src="upload/${meetingVO.meetingSaveImg}" alt="" />
+						<div>
+							<div>일시 <fmt:formatDate value="${meetingDate}" pattern="MM/dd (E) aa hh:mm "/></div>
+							<div>위치 ${meetingVO.meetingLoc}</div>
+							<div>비용 ${meetingVO.meetingPrice}원</div>
+							<div>참석 ${meetingVO.meetingAttend}/${meetingVO.meetingMax} (${meetingVO.meetingMax-meetingVO.meetingAttend}명남음) </div>
+						</div>
 					</div>
 				</div>
-			</div>
-			
-			<div>
-			<!-- 박스 밑의 사용자 이미지아이콘 띄우기 -->
-			<c:forEach var="meetingMemberVO" items="${meetingMemberList}">
+				
+				<div >
 					<span>
-						<c:if test="${meetingMemberVO.meetingNo == meetingVO.meetingNo}" >
-							<a href="" style="text-decoration: none;"><span>
-								<c:if test="${empty meetingMemberVO.memberSaveImg}">
-									<box-icon type='solid' name='user-circle' style="margin-right: 5px"></box-icon>
+						<c:if test="${meetingMemberList[status.index].meetingNo eq meetingVO.meetingNo}" >
+							<a href="" style="text-decoration: none;"><div>
+								<c:if test="${empty meetingMemberList[status.index].memberSaveImg}">
+									<box-icon type='solid' name='user-circle'></box-icon>
 								</c:if>
-								<c:if test="${not empty meetingMemberVO.memberSaveImg}">
-									<img class="crew-member-img" src="upload/${meetingMemberVO.memberSaveImg}" style="margin-right: 5px" />
+								<c:if test="${not empty meetingMemberList[status.index].memberSaveImg}">
+									<img class="crew-member-img" src="upload/${meetingMemberList[status.index].memberSaveImg}" />
 								</c:if>
-							</span></a>
+							</div></a>
 						</c:if>
 					</span>
-			</c:forEach>
-			</div>
+				</div>
 		</c:forEach>
-		
 		</div>
 	</div>
-	
-	
-	<!--모달 팝업-->
-	<%-- <div class="modal">
-	    <div class="modal_popup">
-	        <h3>멤버관리</h3>
-	        <c:forEach var="crewMemberVO" items="${crewMemberList}">
-			<c:if test="${crewMemberVO.crewMemberStatus != 4}">
-				<a href="" style="text-decoration: none;"><div class="crew-member-item">
-					<c:if test="${empty crewMemberVO.memberSaveImg}">
-						<box-icon type='solid' name='user-circle'></box-icon>
-					</c:if>
-					<c:if test="${not empty crewMemberVO.memberSaveImg}">
-						<img class="crew-member-img" src="upload/${crewMemberVO.memberSaveImg}" alt="${crewMemberVO.memberNickname} 프로필 이미지" />
-					</c:if>
-					<span class="crew-member-nickname">${crewMemberVO.memberNickname}</span>
-					<c:if test="${crewMemberVO.crewMemberStatus == 4}">
-						<a href="">승인</a>
-						<a href="">거절</a>
- 					</c:if>
-				</div></a>
-			</c:if>
-			</c:forEach>
-	        <button type="button" class="close_btn">닫기</button>
-	    </div>
-	</div> --%>
 	<jsp:include page="../component/footer.jsp"></jsp:include>
 </div>
 <script>
+// 상단 팝업
 document.addEventListener('DOMContentLoaded', function() {
     var infoIcon = document.querySelector('.crew-info-icon');
     var popup = document.querySelector('.edit-delete-popup');
-    
-    /* var memberManagementWrap = document.querySelector('.member-management-wrap');
-    var modal = document.querySelector('.modal');
-    var closeBtn = document.querySelector('.close_btn'); */
     
     // 팝업
     infoIcon.addEventListener('click', function(event) {
@@ -491,31 +423,61 @@ document.addEventListener('DOMContentLoaded', function() {
             popup.style.display = 'none';
         }
     });
-    
-    
- 	/* // 모달 띄우기
-    memberManagementWrap.addEventListener('click', function(event) {
-        modal.style.display = 'block';
-        event.stopPropagation(); // 이벤트 전파 막기
+});
+
+<!-- SweetAlert2 -->
+$(document).ready(function () {
+	/* 기본 */
+	$('.crew-alert').click(function (event) {
+        event.preventDefault(); 
+        var href = $(this).attr('href'); 
+        var message = $(this).data('message'); 
+
+        Swal.fire({
+            title: "Meetup",
+            text: message,
+            icon: 'success',
+            timer: 1000, 			// 1초 후에 자동으로 닫히도록 설정
+            timerProgressBar: true, // 타이머 진행 바 표시
+            didOpen: () => {
+                Swal.showLoading()
+            },
+        }).then(() => {
+        	window.location.href = href;
+        });
+        
+        // 1초 후에 페이지 이동
+        setTimeout(() => {
+            window.location.href = href;
+        }, 1000);
     });
+	
+	/* confirm */
+    $('.crew-alert-confirm').click(function (event) {
+        event.preventDefault(); 
+        var href = $(this).attr('href'); 
+        var message = $(this).data('message'); 
 
-    // 모달 닫기
-    closeBtn.addEventListener('click', function(event) {
-        modal.style.display = 'none';
+        Swal.fire({
+            title: "Meetup",
+            text: message,
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: '확인',
+            cancelButtonText: '취소'
+        }).then((result) => {
+            if (result.isConfirmed) {
+            		Swal.fire(
+                  'Meetup',
+                  '처리되었습니다.',
+                  	'success'
+                )
+                window.location.href = href; 
+            }
+        });
     });
-
-    // 클릭 시 모달을 숨기기 위한 이벤트 리스너 추가
-    document.addEventListener('click', function(event) {
-        if (!popup.contains(event.target) && event.target !== infoIcon) {
-            popup.style.display = 'none';
-        }
-    });
-
-    // 모달 클릭 시 이벤트 전파 막기
-    modal.addEventListener('click', function(event) {
-        event.stopPropagation();
-    }); */
-
 });
 </script>
 </body>
